Add Annotation to P0wnedPassword
Without this change i become this error:

The class "Rollerworks\Component\PasswordStrength\Validator\Constraints\P0wnedPassword" is not annotated with @Annotation. Are you sure this class can be used as annotation? If so, then you need to add @Annotation to the _class_ doc comment of "Rollerworks\Component\PasswordStrength\Validator\Constraints\P0wnedPassword".